Fixed calcElutionTime() Bug
Elution time calculation method was off. Sometimes, the amount of simulation steps remaining would be a double that was larger than integer.MAX_VALUE. When any double larger than Interger.MAX_VALUE is cast to an int, it becomes Iteger.MAX_VALUE. This was leading to the elution times of any peaks with sufficiently long elution times (temps low enough) to always have the same elution time: Integer.MAX_VALUE * 0.05 (FRAME_LENGTH_S = 0.05)

+++ src/main/java/com/karimbouchareb/chromatographysimulator/ChromatographySimulator.java
@@ -904,8 +904,8 @@ public class ChromatographySimulator extends Application {
         private double proportionOfColumnTraversed(){
             return proportionOfColumnTraversed;
         }
-        private int simulationStepsRemainingUntilPeakElutes(){
-            return (int) Math.ceil(proportionOfColumnUntraversed()
+        private double simulationStepsRemainingUntilPeakElutes(){
+            return Math.ceil(proportionOfColumnUntraversed()
                     /traversalProgressPerSimulationStep());
         }
         private double secondsRemainingUntilPeakElutes(){
